Add initial state with link to estimate shipping
