+++ CHANGELOG.md
@@ -7,15 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Initial state with link to show delivery options.
+
 ## [0.1.1] - 2019-08-21
 
 ## Fixed
 
-- Container styles;
+- Container styles.
 
 ## Added
 
-- Default props with new address if it does not come from props
+- Default props with new address if it does not come from props.
 
 ## [0.1.0] - 2019-08-20
