+++ messages/es.json
@@ -2,5 +2,6 @@
   "store/shipping-calculator.delivery": "Entrega",
   "store/shipping-calculator.deliveryFor": "Entrega en",
   "store/shipping-calculator.edit": "cambiar",
-  "store/shipping-calculator.estimate": "calcular"
+  "store/shipping-calculator.estimate": "calcular",
+  "store/shipping-calculator.viewDeliveryOptions": "Ver opciones de entrega"
 }
