+++ messages/pt.json
@@ -2,5 +2,6 @@
   "store/shipping-calculator.delivery": "Entrega",
   "store/shipping-calculator.deliveryFor": "Entrega em",
   "store/shipping-calculator.edit": "editar",
-  "store/shipping-calculator.estimate": "calcular"
+  "store/shipping-calculator.estimate": "calcular",
+  "store/shipping-calculator.viewDeliveryOptions": "Ver opções de entrega"
 }
